RoomCard wurde in mehrere Komponenten aufgeteilt
